% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_digit.R
\name{check_digit}
\alias{check_digit}
\title{Calc check digits of municipality codes}
\usage{
check_digit(code)
}
\arguments{
\item{code}{A character vector; municipality codes.}
}
\value{
A character vector.
}
\description{
Calc check digits of municipality codes
}
