% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jisx0402-package.R
\docType{package}
\name{jisx0402-package}
\alias{jisx0402}
\alias{jisx0402-package}
\title{jisx0402: Datasets Related to 'JIS X 0402:2020'}
\description{
Provides datasets for handling Japanese municipality code defined in 'JIS X 0402' and 'JIS X 0401'.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/paithiov909/jisx0402/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
