% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jisx0402-package.R
\docType{data}
\name{jpaddresses}
\alias{jpaddresses}
\title{Japanese addresses and their coordinates}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 277656 rows and 6 columns.
}
\source{
\url{https://github.com/geolonia/japanese-addresses/raw/develop/data/latest.csv}
}
\usage{
jpaddresses
}
\description{
Modified from the 'latest.csv' file
in \href{https://github.com/geolonia/japanese-addresses/}{japanese-addresses}.
The original dataset is licensed under \href{https://creativecommons.org/licenses/by/4.0/deed.ja}{CC BY 4.0}.
}
\keyword{datasets}
