% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jisx0402-package.R
\docType{data}
\name{jpprefs}
\alias{jpprefs}
\title{Prefecture codes of Japan}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 47 rows and 2 columns.
}
\usage{
jpprefs
}
\description{
Prefecture codes of Japan
}
\keyword{datasets}
