% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jptopography.R
\name{jptopography}
\alias{jptopography}
\title{Topographic data of Japanese municipalities}
\usage{
jptopography(type = c("all", "designated", "prefecture"), resolution = 0.1)
}
\arguments{
\item{type}{String; One of \code{all}, \code{desiginated}, or \code{prefecture}.}

\item{resolution}{Numeric; If smaller than \code{1},
returns the 0.1\% desolved data ('s0001').
Otherwise, returns the 1\% desolved data ('s0010').}
}
\value{
An \code{sf} object from the 'sf' package.
}
\description{
Reads 'FlatGeobuf' file and returns topographic data as an \code{sf} object.
}
\details{
This package contains several 'FlatGeobuf' files converted
from the 'GeoJSON' files of \href{https://github.com/smartnews-smri/japan-topography}{smartnews-smri/japan-topography}.
The original 'GeoJSON' files are sourced from
\href{https://nlftp.mlit.go.jp/ksj/}{MLIT of Japan site}.
Therefore, to use these dataset,
you must agree to their \href{https://nlftp.mlit.go.jp/ksj/other/agreement.html}{term of use}.
}
