% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jisx0402-package.R
\docType{data}
\name{municipality}
\alias{municipality}
\title{Municipality codes of Japan}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4542 rows and 6 columns.
}
\usage{
municipality
}
\description{
A list of Japanese municipalities and their codes,
as known as the 'Zenkoku Chiho-kokyo-dantai Code'.
You can use a concatnated string like: \code{paste0(pref_code, city_code)}
as an actual municipality code.
}
\details{
The 'Zenkoku Chiho-kokyo-dantai Code' is a combination
of a prefecture code defined in 'JIS X 0401'
and a municipality code (city, ward, town or village code) defined in 'JIS X 0402'.

These codes consists of 5 to 6 digits.
The 6th digit is optional; that is a check digit.
Of the remaining numbers, the first 2 digits are the prefecture code
and the last 3 digits are the municipality code.

'000' as a municipality code means prefecture itself.
For example, '010006' is a hybrid of '01' ('pref_code' of Hokkaido),
'000', and its check digit '6'.
}
\seealso{
\url{https://www.soumu.go.jp/denshijiti/code.html}
}
\keyword{datasets}
